Fix: rename util modules
