+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix: Parse up to 100 issues in a request (@lwasser, #94)
 - Fix: refactor parse issue header (@lwasser, #91)
 - Fix: Refactor and organize modules (@lwasser, #76)
+- Fix: Rename and organize `clean.py` module into `utils_parse` and `utils_clean` (@lwasser, @willingc, #121)
 
 
 ## [v0.2.3] - 2024-02-29
