+++ src/pyosmeta/utils_parse.py
@@ -1,0 +1,47 @@
+"""
+A small module containing functions that support parsing
+pyOpenSci review and contributor metadata.
+"""
+
+from pyosmeta.utils_clean import clean_name
+
+
+def parse_user_names(username: str) -> dict[str, str]:
+    """Parses authors, contributors, editors and usernames from
+    the requested issues.
+
+    Parameters
+    ----------
+    username : str
+        The line with username (see Notes).
+
+    Returns
+    -------
+    dict
+        ``{name: str, github_username: str}``
+
+    Notes
+    -----
+    Possible combinations:
+
+    1. Name Surname (@Github_handle)
+    2. (@Github_handle)
+    3. @Github_handle
+
+    """
+    names = username.split("@", 1)
+    names = [x for x in names if len(x) > 0]
+
+    if len(names) > 1:
+        parsed = {
+            "github_username": clean_name(names[1]),
+            "name": clean_name(names[0]),
+        }
+
+    else:
+        parsed = {
+            "github_username": clean_name(names[0]),
+            "name": "",
+        }
+
+    return parsed
